// api/types/paths/notifications/v1/destinations.types.ts
export type HTTP_GET = (
  $: OmitValueWhenNever<{
    query: never;
    path: never;
    header: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetDestinationsResponse;
          };
        };
      };
      400: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetDestinationsResponse;
          };
        };
      };
      403: {
        headers: {
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetDestinationsResponse;
          };
        };
      };
      404: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetDestinationsResponse;
          };
        };
      };
      409: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetDestinationsResponse;
          };
        };
      };
      413: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetDestinationsResponse;
          };
        };
      };
      415: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetDestinationsResponse;
          };
        };
      };
      429: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetDestinationsResponse;
          };
        };
      };
      500: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetDestinationsResponse;
          };
        };
      };
      503: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetDestinationsResponse;
          };
        };
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
  }>
) =>
  | {
      status: 200;
      contentType?: 'application/json';
      body?: GetDestinationsResponse;
    }
  | {
      status: 400;
      contentType?: 'application/json';
      body?: GetDestinationsResponse;
    }
  | {
      status: 403;
      contentType?: 'application/json';
      body?: GetDestinationsResponse;
    }
  | {
      status: 404;
      contentType?: 'application/json';
      body?: GetDestinationsResponse;
    }
  | {
      status: 409;
      contentType?: 'application/json';
      body?: GetDestinationsResponse;
    }
  | {
      status: 413;
      contentType?: 'application/json';
      body?: GetDestinationsResponse;
    }
  | {
      status: 415;
      contentType?: 'application/json';
      body?: GetDestinationsResponse;
    }
  | {
      status: 429;
      contentType?: 'application/json';
      body?: GetDestinationsResponse;
    }
  | {
      status: 500;
      contentType?: 'application/json';
      body?: GetDestinationsResponse;
    }
  | {
      status: 503;
      contentType?: 'application/json';
      body?: GetDestinationsResponse;
    }
  | { status: 415; contentType: 'text/plain'; body: string }
  | COUNTERFACT_RESPONSE
  | { ALL_REMAINING_HEADERS_ARE_OPTIONAL: COUNTERFACT_RESPONSE };

export type HTTP_POST = (
  $: OmitValueWhenNever<{
    query: never;
    path: never;
    header: never;
    body: CreateDestinationRequest;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: CreateDestinationResponse;
          };
        };
      };
      400: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: CreateDestinationResponse;
          };
        };
      };
      403: {
        headers: {
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: CreateDestinationResponse;
          };
        };
      };
      404: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: CreateDestinationResponse;
          };
        };
      };
      409: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: CreateDestinationResponse;
          };
        };
      };
      413: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: CreateDestinationResponse;
          };
        };
      };
      415: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: CreateDestinationResponse;
          };
        };
      };
      429: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: CreateDestinationResponse;
          };
        };
      };
      500: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: CreateDestinationResponse;
          };
        };
      };
      503: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: CreateDestinationResponse;
          };
        };
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
  }>
) =>
  | {
      status: 200;
      contentType?: 'application/json';
      body?: CreateDestinationResponse;
    }
  | {
      status: 400;
      contentType?: 'application/json';
      body?: CreateDestinationResponse;
    }
  | {
      status: 403;
      contentType?: 'application/json';
      body?: CreateDestinationResponse;
    }
  | {
      status: 404;
      contentType?: 'application/json';
      body?: CreateDestinationResponse;
    }
  | {
      status: 409;
      contentType?: 'application/json';
      body?: CreateDestinationResponse;
    }
  | {
      status: 413;
      contentType?: 'application/json';
      body?: CreateDestinationResponse;
    }
  | {
      status: 415;
      contentType?: 'application/json';
      body?: CreateDestinationResponse;
    }
  | {
      status: 429;
      contentType?: 'application/json';
      body?: CreateDestinationResponse;
    }
  | {
      status: 500;
      contentType?: 'application/json';
      body?: CreateDestinationResponse;
    }
  | {
      status: 503;
      contentType?: 'application/json';
      body?: CreateDestinationResponse;
    }
  | { status: 415; contentType: 'text/plain'; body: string }
  | COUNTERFACT_RESPONSE
  | { ALL_REMAINING_HEADERS_ARE_OPTIONAL: COUNTERFACT_RESPONSE };